+++ sdk/monitor/azure-monitor-query/tests/test_logs_client.py
@@ -28,6 +28,18 @@ def test_logs_single_query():
     assert response.tables is not None
 
 @pytest.mark.live_test_only
+def test_logs_single_query_raises_no_timespan():
+    credential = _credential()
+    client = LogsQueryClient(credential)
+    query = """AppRequests | 
+    where TimeGenerated > ago(12h) | 
+    summarize avgRequestDuration=avg(DurationMs) by bin(TimeGenerated, 10m), _ResourceId"""
+
+    # returns LogsQueryResult 
+    with pytest.raises(TypeError):
+        client.query(os.environ['LOG_WORKSPACE_ID'], query)
+
+@pytest.mark.live_test_only
 def test_logs_single_query_with_non_200():
     credential = _credential()
     client = LogsQueryClient(credential)